add validate password

+++ accounts/serializers.py
@@ -13,7 +13,14 @@ class UserRegisterSerializer(serializers.ModelSerializer):
             'password': {'write_only': True}
         }
 
+    def validate_password(self, value):
+        if len(value) < 8:
+            raise serializers.ValidationError('Password must be at least 8 characters')
+
     def validate(self, attrs):
+        if attrs['password'] != attrs['c_password']:
+            raise serializers.ValidationError("Passwords do not match")
+
         user = User.objects.filter(email=attrs['email']).first()
         if user and not user.is_active:
             user.delete()
